+++ src/fs/write_file.rs
@@ -2,7 +2,8 @@ use rquickjs::{Ctx, Result, Value};
 use tokio::fs;
 use tokio::io::AsyncWriteExt;
 
-use crate::util::{get_bytes, ResultExt};
+use crate::utils::object::get_bytes;
+use crate::utils::result::ResultExt;
 
 pub async fn write_file<'js>(ctx: Ctx<'js>, path: String, data: Value<'js>) -> Result<()> {
     let mut file = fs::File::create(&path)