+++ src/module.rs
@@ -1,10 +1,26 @@
 use rquickjs::{
     module::{Declarations, Exports, ModuleDef},
     prelude::Func,
-    Ctx, Result, Value,
+    Ctx, Object, Result, Value,
 };
 
-use crate::util::export_default;
+pub fn export_default<'js, F>(ctx: &Ctx<'js>, exports: &mut Exports<'js>, f: F) -> Result<()>
+where
+    F: FnOnce(&Object<'js>) -> Result<()>,
+{
+    let default = Object::new(ctx.clone())?;
+    f(&default)?;
+
+    for name in default.keys::<String>() {
+        let name = name?;
+        let value: Value = default.get(&name)?;
+        exports.export(name, value)?;
+    }
+
+    exports.export("default", default)?;
+
+    Ok(())
+}
 
 pub struct ModuleModule;