+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 * Switch libc from gnu to musl
 * Bug fixes
 * Dependency upgrades
+* Implement `structuredClone` in global object